+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added `-k/--insecure` flag to disable SSL certificate verification
+- Added `json` output feature to analyze module
 
 ### Changed